requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
jinja2==3.1.2
apscheduler==3.10.4
websockets==12.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psutil==5.9.6
aiohttp==3.9.1
requests==2.31.0
Pillow==10.1.0
